refactor: dto package 구조 리팩토링
