+++ src/main/java/inno/escape/moviebackend/domain/boxoffice/dto/daily/DailyBoxOfficeRequestDto.java
@@ -1,4 +1,4 @@
-package inno.escape.moviebackend.domain.boxoffice.dto;
+package inno.escape.moviebackend.domain.boxoffice.dto.daily;
 
 import java.io.Serializable;
 import javax.validation.constraints.NotBlank;
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Data
-public class WeeklyBoxOfficeRequestDto implements Serializable {
+public class DailyBoxOfficeRequestDto implements Serializable {
 
   @NotBlank(message = "발급받은 키 값을 입력해주세요.")
   private String key;
@@ -21,9 +21,6 @@ public class WeeklyBoxOfficeRequestDto implements Serializable {
   @Pattern(regexp = "((19|20)\\d{2}(0[1-9]|1[0-2])(0[1-9]|[12]\\d|3[01]))", message = "조회하고자 하는 날짜를 yyyymmdd 형식으로 입력해주세요.")
   @NotBlank(message = "조회하고자 하는 날짜를 입력해주세요.")
   private String targetDt;
-
-  // 0 주간(월~일) / default 1 주말(금~일) / 2 주중(월~목)
-  private String weekGb;
 
   // default 10 / max 10
   private String itemPerPage;
@@ -34,6 +31,7 @@ public class WeeklyBoxOfficeRequestDto implements Serializable {
   // default 전체 / K 한국 영화 / F 외국 영화
   private String repNationCd;
 
-  // default 전체 / 0105001 서울시 ~ 0105017 세종시
+  // default 전체
+  // 공통코드 조회 서비스에서 "0105000000" 로서 조회된 코드 => 0105001 서울시 ~ 0105017 세종시
   private String wideAreaCd;
 }
