+++ src/main/java/inno/escape/moviebackend/domain/boxoffice/dto/daily/DailyBoxOfficeResponseDto.java
@@ -1,4 +1,4 @@
-package inno.escape.moviebackend.domain.boxoffice.dto;
+package inno.escape.moviebackend.domain.boxoffice.dto.daily;
 
 import java.io.Serializable;
 import lombok.AccessLevel;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Data
-public class WeeklyBoxOfficeResponseDto implements Serializable {
+public class DailyBoxOfficeResponseDto implements Serializable {
 
-  private WeeklyBoxOfficeResult boxOfficeResult;
+  private DailyBoxOfficeResult boxOfficeResult;
 }
