+++ src/main/java/inno/escape/moviebackend/domain/boxoffice/service/BoxOfficeService.java
@@ -1,9 +1,9 @@
 package inno.escape.moviebackend.domain.boxoffice.service;
 
-import inno.escape.moviebackend.domain.boxoffice.dto.DailyBoxOfficeRequestDto;
-import inno.escape.moviebackend.domain.boxoffice.dto.DailyBoxOfficeResponseDto;
-import inno.escape.moviebackend.domain.boxoffice.dto.WeeklyBoxOfficeRequestDto;
-import inno.escape.moviebackend.domain.boxoffice.dto.WeeklyBoxOfficeResponseDto;
+import inno.escape.moviebackend.domain.boxoffice.dto.daily.DailyBoxOfficeRequestDto;
+import inno.escape.moviebackend.domain.boxoffice.dto.daily.DailyBoxOfficeResponseDto;
+import inno.escape.moviebackend.domain.boxoffice.dto.weekly.WeeklyBoxOfficeRequestDto;
+import inno.escape.moviebackend.domain.boxoffice.dto.weekly.WeeklyBoxOfficeResponseDto;
 import reactor.core.publisher.Mono;
 
 public interface BoxOfficeService {
